+++ src/Entity/Interfaces/StatusHistoryContainer.php
@@ -4,11 +4,29 @@ namespace App\Entity\Interfaces;
 
 use App\Entity\StatusHistory;
 use App\Entity\Statut;
+use DateTime;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\Collections\Criteria;
+use Doctrine\ORM\Mapping as ORM;
+use Doctrine\DBAL\Types\Types;
 use WiiCommon\Helper\Stream;
 
+#[ORM\MappedSuperclass()]
 abstract class StatusHistoryContainer {
+    #[ORM\Column(type: Types::DATETIME_MUTABLE, nullable: true)]
+    private ?DateTime $lastPartialStatusDate = null;
+
+    public function getLastPartialStatusDate(): ?DateTime
+    {
+        return $this->lastPartialStatusDate;
+    }
+
+    public function setLastPartialStatusDate(?DateTime $lastPartialStatusDate): self
+    {
+        $this->lastPartialStatusDate = $lastPartialStatusDate;
+
+        return $this;
+    }
 
     /**
      * @return Collection<int, StatusHistory>
